% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_data_sets}
\alias{available_data_sets}
\title{Available data sets.}
\format{A data frame with 270 rows and 2 variables:
\describe{
  \item{dataset.value"}{URI}
  \item{subject}{Shortened name}
}}
\source{
\url{https://statistics.gov.scot/}
}
\usage{
available_data_sets
}
\description{
A dataset containing the data set names and URIs disseminated via
  \href{https://statistics.gov.scot/}{Statistics.gov.scot}.
}
\keyword{datasets}
