% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_endpoint.R
\name{check_endpoint}
\alias{check_endpoint}
\title{Check SPARQL endpoint}
\usage{
check_endpoint(check_mode = c("warn", "stop", "ignore"), timeout = 1)
}
\arguments{
\item{check_mode}{For \code{stop, warn} or \code{ignore}, decides whether to
stop, warn or do nothing in case of broken endpoint url. \code{ignore}
option ignores checking.}

\item{timeout}{Timeout for a ping response,
as in \code{\link[pingr]{timeout}}}

\item{endpoint}{A SPARQL enpoint as string.}
}
\value{
A logical.
}
\description{
Function checks whether the provided endpoint is of online.
}
\examples{
\dontrun{
  check_endpoint("http://statistics.gov.scot/sparql")
}
}
