% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_filter.R
\name{construct_filter}
\alias{construct_filter}
\title{Construct SPARQL filter}
\usage{
construct_filter(sparql_variable, filter_values, use_str = FALSE)
}
\arguments{
\item{sparql_variable}{A character scalar to be used in \code{IN} clause,
without ?.}

\item{filter_values}{A vector of filter values to be verified against in the
filter statement.}

\item{use_str}{A logical, defaults to \code{FALSE}, if \code{TRUE} it
produces \code{str(?variable)} instead of \code{?variable}.}
}
\value{
A character scalar corresponding to SPARQL filter expression.
}
\description{
This query derives \code{SPARQL} filter statement using the
  \code{IN} syntax.
}
\details{
For instance, arguments \code{sparql_variable = "time"} and
  \code{filter_values = c(2011, 2012)} will result in a expression
  \code{FILTER (?time IN ( "2011, 2012" ))}.
}
\examples{
# This function is not exported from the package as it's mostly used
# internally.
SmarterScotland:::construct_filter(sparql_variable = "time", filter_values = c(2011, 2012))
}
