% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_geography_URI.R
\name{find_geography_URI}
\alias{find_geography_URI}
\title{Match Geography URI}
\usage{
find_geography_URI(geography, database = "internal")
}
\arguments{
\item{geography}{A character vector of geography names. URIs starting with
\code{http} are returned as such without searching the data set.}

\item{database}{Can be one of the following \emph{internal, ScotStat, both.}
Defaults to \emph{internal} that reflects an extract of Standard Geography
Code Register. Option \emph{online} searches across the
\href{https://statistics.gov.scot/}{Statistics.gov.scot}.}
}
\value{
A named character vector of unique geography URIs. For multiple
  matches names will be numbered - string \emph{'Edinburgh'} will
  match multiple geographies that will be named
  \emph{'Edinburgh1', 'Edinburgh2', ...}
}
\description{
Matches passed string with URI available in
  the \emph{Standard Geography Code Register}.For URI's function returned
  passed values.
}
\details{
The function is mostly used internally to make using
  \code{\link[SmarterScotland]{get_geography_data}} more user friendly. The
  function can be used in isolation to conveniently get URIs for the
  colloquial geography names.
}
\examples{
find_geography_URI(geography = c("Edinburgh", "Glasgow"))
}
