% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_geographies.R
\name{get_available_geographies}
\alias{get_available_geographies}
\title{Get Available Geographies}
\usage{
get_available_geographies()
}
\value{
A data frame.
}
\description{
The function sources available goegraphies from
  \emph{Standard Geography Code Register}.
}
\details{
The \emph{Standard Geography Code Register} is available in the
  package and can be loaded using \code{data(available_geographies)} command.
  This function can be used to refresh this data set. The geographies data
  is not updated freauently and is fairly big so in majority of cases it is
  likely that users will find it easier to use the data set available in the
  package.
}
\examples{
\dontrun{
get_available_geographies()
}
}
