% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_hierarchies.R
\name{get_available_hierarchies}
\alias{get_available_hierarchies}
\title{Get Available Spatial Frameworks}
\usage{
get_available_hierarchies(endpoint = "http://statistics.gov.scot/sparql.csv")
}
\value{
A data frame containing list of available hierarchies.
}
\description{
The following function takes no arguments other than the optional
  endpoint and sources available spatial hierarchies. The returned hierarchies
  correspond to the available best-fit geographies and can be used to aggregate
  data zone level data.
}
\details{
In practice, the function can be used to to inform aguments in
  \code{\link{get_geographies_data}} and in
  \code{\link{get_available_geographies}}.
}
\examples{
\dontrun{
head(get_available_hierarchies())
#   hierarchy                                                           label
# 1 http://statistics.gov.scot/def/hierarchy/best-fit#health-board-area Best fit health board area
# 2      http://statistics.gov.scot/def/hierarchy/best-fit#council-area Best fit council area
}
}
