% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(indicator, geography)
}
\arguments{
\item{indicator}{A character vector with the name of indicator to be sourced from the
\href{http://statistics.gov.scot}{statistics.gov.scot}.}
}
\description{
This is a workhorse function of this package and it's used to
  source a data frame from the \href{http://statistics.gov.scot}{statistics.gov.scot}.
}
\details{
In effect, the following function works as a wraper for a SPARQL query
  sourcing data across desired geographies. The query requires a number of dimensions
  concerned with the indicator, value and geography to be specified.
}
