% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_properties.R
\name{get_data_properties}
\alias{get_data_properties}
\title{Get Data Properties}
\usage{
get_data_properties(data_set, property_values = TRUE)
}
\arguments{
\item{data_set}{String corresponding to \code{subject} column derived via
\code{\link[SmarterScotland]{get_available_datasets}}.}

\item{property_values}{A boolean. Defults to \code{FALSE}. If \code{TRUE}
the function will return values for each available property.}
}
\value{
A data frame.
}
\description{
The queries sources properties for the desired data set and
  optionally values for those properties.
}
\examples{
\dontrun{
get_data_properties(data_set = "life-expectancy")
get_data_properties(data_set = "life-expectancy", property_values = TRUE)
}
}
