% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geography_code_register.R
\name{get_geography_code_register}
\alias{get_geography_code_register}
\title{Get Standard Geography Code Register}
\usage{
get_geography_code_register()
}
\value{
A data frame.
}
\description{
The function sources \emph{Standard Geography Code Register}.
}
\details{
The \emph{Standard Geography Code Register} is available in the
  package and can be loaded using
}
\examples{
\dontrun{
dta_geo <- get_geography_code_register()
}
}
