% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geography_data.R
\name{get_geography_data}
\alias{get_geography_data}
\title{Get Geography Data}
\usage{
get_geography_data(data_set, geography, period)
}
\arguments{
\item{data_set}{\strong{Required.} A character vector with the name of
indicator to be sourced from the
\href{http://statistics.gov.scot}{statistics.gov.scot}.}

\item{geography}{\strong{Required.} A character vector corresponding to
geographies for which to derive the data this can be a common}

\item{period}{\strong{Optional.} A character vector corresponding to period(s)
for which the data should be sourced. This has to correspond to the
available periods. For instance \code{c("2001Q2", "2010Q2")} would get the
data for 2nd quarters of 2001 and 2010 respectively.}
}
\value{
A data frame.
}
\description{
The function fetches statistical data for a selected geography
  framework from: \href{http://statistics.gov.scot}{statistics.gov.scot}
  for a selected period for a chosen indicator.
}
\details{
The following function attempts to source all data
  available for a specific geographies for the provided period. The period
  and geography names are used in SPARQL via \code{FILTER} clause where
  vector values are used to construct and if statement. Those values are
  optional but it's recommended that are those are recommend as without those
  the query would attempt to return all data set values for all period for
  all geographies.
}
\section{Available Geographies and Data}{

Helper functions and static data sets useful in identifying available
 geographies and data sets are provided in this package.
}

\examples{
\dontrun{
get_geography_data(data_set = "alcohol-related-discharge",
                   geography = "datazones")
}
}
