% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geography_hierarchies.R
\name{get_geography_hierarchies}
\alias{get_geography_hierarchies}
\title{Get Available Spatial Frameworks}
\usage{
get_geography_hierarchies()
}
\value{
A data frame containing list of available hierarchies.
}
\description{
The following function takes no arguments other than the optional
  endpoint and sources available spatial hierarchies. The returned hierarchies
  correspond to the available best-fit geographies and can be used to aggregate
  data zone level data.
}
\details{
In practice, the function can be used to to inform arguments in
  \code{\link[SmarterScotland]{get_geography_data}}.
}
\examples{
\dontrun{
head(get_available_hierarchies())
#   hierarchy                                                           label
# 1 http://statistics.gov.scot/def/hierarchy/best-fit#health-board-area Best fit health board area
# 2      http://statistics.gov.scot/def/hierarchy/best-fit#council-area Best fit council area
}
}
