% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_data.R
\name{pre_process_data}
\alias{pre_process_data}
\title{Pre-Process SPARQL Results Data}
\usage{
pre_process_data(x, clean_URI_strings = TRUE, remove_cols = TRUE)
}
\arguments{
\item{x}{A data frame, usually obtained via
\code{\link[SmarterScotland]{get_geography_data}} or other data sourcing
function.}

\item{clean_URI_strings}{Defaults to \code{TRUE}; removes initial part of
URI \code{ex. http://purl.org/linked-data/cube#} from an URI string.}

\item{remove_cols}{Removes redundant columns, such as columns with value
\code{URI} only. Defaults to \code{TRUE}.}
}
\value{
A data frame.
}
\description{
The function takes data frame generated by one of the functions
  that sourced data from \href{http://statistics.gov.scot}{statistics.gov.scot}
  and returns pre-processed data set with the following changes.
}
\examples{
\dontrun{
pre_process_data(x = get_geography_data(data_set = "recorded-crime",
                                        geography = "Glasgow City",
                                        measure = "count"))
}
}
