% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_scotstat.R
\name{query_scotstat}
\alias{query_scotstat}
\title{Query Scotstat Repository}
\usage{
query_scotstat(query)
}
\arguments{
\item{A}{character scalar corresponding to SPARQL query string.}
}
\value{
A response object derived via \code{\link[httr]{POST}} function.
}
\description{
Send query to statistics.gov.scot.
}
\details{
Query Scotstat data using \code{\link[httr]{POST}} function
 available in \code{\link[httr]{httr}} package. The functions reads the 
 \code{SmarterScotland.endpooint} option values and uses the provided
 endpoint to send query to statistics.gov.scot.
}
\examples{
\dontrun{query_scotstat(read_query_file(query_file("get_available_frameworks")))}


}
