% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmarterScotland.R
\docType{package}
\name{SmarterScotland-package}
\alias{SmarterScotland}
\alias{SmarterScotland-package}
\title{Sourcing data from opendata.gov.scot}
\description{
SmarterScotland provides a suite of conveniance functions
  facilitating sourcing data from the \href{http://statistics.gov.scot}{statistics.gov.scot}
  portal.
}
\details{
For sample usage see \code{vignette("basic-usage", package = "SmarterScotland")}.
  For other usage, inluding charting and more advanced functionalities see
 \code{vignette("other-functions", package = "SmarterScotland")}.
}
