% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_geography_code_register}
\alias{standard_geography_code_register}
\title{Standard Geography Code Register}
\format{A data frame with three columns:
\describe{
  \item{geography}{Geography URI}
  \item{property}{Geography Property such as label or official name}
  \item{value}{Property value}
}}
\source{
\url{https://statistics.gov.scot/}
}
\usage{
standard_geography_code_register
}
\description{
The following data set contains trimmed version of the
  \emph{Standard Geography Code Register}. The full data set is available at
  dedicated URL
  \url{https://statistics.gov.scot/data/standard-geography-code-register}.
}
\keyword{datasets}
