% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{string_utilities}
\alias{string_utilities}
\alias{make_SPARQL_variable_names}
\alias{create_URI}
\title{Query File Utilities}
\usage{
make_SPARQL_variable_names(x)

create_URI(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Utility functions to handle frequent string operations.
}
\details{
The functions take character vector and are used to accomplish
  simple string operations, such as replacement of undesired characters
  or other simple transformations. The functions are not exported as it's
  unlikely that anyone would found them useful outside the package
  environment.
}
\section{make_SPARQL_variable_names}{
 Wrapper for the
  \code{\link[base]{gsub}} function that replaces special characters with
  underscore. Collections of special characters are replaced with one
  underscore character.

 Wrapper for the
  \code{\link[base]{gsub}} function that replaces special characters with
  underscore. Collections of special characters are replaced with one
  underscore character. By default, selected strings will be suffixed with
  \code{/rdfs:label}. This behaviour can be controlled via
  \code{SmarterScotland.suffix_rdf} option.
}

