% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_package_version.R
\name{update_package_version}
\alias{update_package_version}
\title{Update Package Version}
\usage{
update_package_version(packageLocation = ".")
}
\value{
Integer
}
\description{
Updates package version
}
\keyword{internal}
