% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/utils.R
\name{read_query_file}
\alias{read_query_file}
\alias{utility_functions}
\alias{query_file}
\title{Handy utility functions}
\usage{
read_query_file(path)

query_file(query_file_name)
}
\description{
Small and handy utility functions usually not useful outside
  the package enviornment.
}
\section{Functions}{
\itemize{
\item \code{read_query_file}: Provides a convenience function quickly reading
\code{SPARQL} files via \code{Rcpp}.

\item \code{query_file}: Wrapper for the \code{system.file} function
passing path to a given query file and taking care of potentially missing
extensions.
}}

\keyword{internal}
