% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_endpoint.R
\name{check_endpoint}
\alias{check_endpoint}
\title{Check SPARQL endpoint}
\usage{
check_endpoint(endpoint, stop_on_bad = FALSE, timeout = 1)
}
\arguments{
\item{endpoint}{A SPARQL enpoint as string.}

\item{stop_on_bad}{A logical, if \code{TRUE} function will
call \code{\link[base]{stop}} if endpoint is not accessible.}

\item{timeout}{Timeout for a ping response,
as in \code{\link[pingr]{timeout}}}
}
\value{
A logical.
}
\description{
Function checks whether the provided endpoint is of online.
}
\examples{
\dontrun{
  check_endpoint("http://statistics.gov.scot/sparql")
}
}
