% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_datasets.R
\name{get_available_datasets}
\alias{get_available_datasets}
\title{Get Available Data Sets}
\usage{
get_available_datasets(pattern,
  endpoint = "http://statistics.gov.scot/sparql.csv")
}
\arguments{
\item{pattern}{A optional pattern to match data set name.}
}
\value{
A data frame
}
\description{
Function sources data available at
  \url{https://statistics.gov.scot/}.
}
\examples{
\dontrun{
get_available_datasets(pattern = "health")
}
}
