% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_geographies.R
\name{get_available_geographies}
\alias{get_available_geographies}
\title{Get Available Geographies}
\usage{
get_available_geographies()
}
\value{
A data frame.
}
\description{
The function sources available goegraphies from
  \emph{Standard Geography Code Register} so it cab be loaded in the package
  using \code{data(available_geographies)} command.
}
\details{
For structure of the data look at
  \code{\link[SmarterScotland]{standard_geography_code_register}}.
}
\examples{
\dontrun{
get_available_geographies()
}
}
