% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geography_data.R
\name{get_geography_data}
\alias{get_geography_data}
\title{Get Geography Data}
\usage{
get_geography_data(indicator, geography, period)
}
\arguments{
\item{indicator}{\strong{Required.} A character vector with the name of
indicator to be sourced from the
\href{http://statistics.gov.scot}{statistics.gov.scot}.}

\item{geography}{\strong{Required.} A character corresponding to geography framework.
The list of available frameworks can be obtained through
\code{\link{get_geography_hierarchies}}.}

\item{period}{\strong{Optional.} A character vector corresponding to period(s)
for which the data should be sourced. This has to correspond to the
available periods. For instance \code{c("2001Q2", "2010Q2")} would get the
data for 2nd quarters of 2001 and 2010 respectively.}
}
\value{
A data frame.
}
\description{
The function fetches statistical data for a selected geography
  framework from: \href{http://statistics.gov.scot}{statistics.gov.scot}
  for a selected period for a chosen indicator.
}
\details{
The following function attempts to source all data
  available for a specific spatial framework for all periods for
  selected spatial hierarchy or only for the selected period if the
  \code{period} argument is provided.
}
\examples{
\dontrun{
get_geography_data(data_set = "alcohol-related-discharge",
                   geography = "datazones")
}
}
