% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geography_data.R
\name{get_geography_data}
\alias{get_geography_data}
\title{Get Geography Data}
\usage{
get_geography_data(data_set, geography, period, measure = "count")
}
\arguments{
\item{data_set}{\strong{Required.} A character vector with the name of
indicator to be sourced from the
\href{http://statistics.gov.scot}{statistics.gov.scot}.}

\item{geography}{\strong{Required.} A character vector corresponding to
geographies for which to derive the data or URI corresponding to a
specific geography. Colloquial}

\item{period}{\strong{Optional.} A character vector corresponding to period(s)
for which the data should be sourced. This has to correspond to the
available periods. For instance \code{c("2001Q2", "2010Q2")} would get the
data for 2nd quarters of 2001 and 2010 respectively.}

\item{measure}{\strong{Required.} A character scalar pertaining to
\emph{count} or \emph{ratio} and
# TODO: Finish this argument}
}
\value{
A data frame.
}
\description{
The function fetches statistical data for a selected geography
  framework from: \href{http://statistics.gov.scot}{statistics.gov.scot}
  for a selected period for a chosen indicator.
}
\details{
The following function attempts to source all data available for
  specific geographies for the provided period periods. Colloquial
  geography names are matched with URI values in
  \emph{Standard Geography Code Register}. A well known locality name like
  \emph{Edinburgh} may match multiple URIs. Have a look at
  \code{\link[SmarterScotland]{available_data_sets}} documentations to
  explore data available in the package and
  \code{\link[SmarterScotland]{get_available_datasets}} function to get most
  recent list.
}
\examples{
\dontrun{
get_geography_data(data_set = "alcohol-related-discharge",
                   geography = "Edinburgh")
}
}
