% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_toc.R
\name{get_ToC}
\alias{get_ToC}
\title{Get TOC from statistics.gov.scot}
\usage{
get_ToC(which = NULL)
}
\arguments{
\item{which}{Defults \code{NULL} returns all subjects, a character value is
passed to \code{FILTER} in SPARQL query performing case insesitive regex
on filter categories.}
}
\value{
A \code{data.frame} with columns corresponding to table name,
  available geographies and time series intervals.
}
\description{
Sources list of data available through statistics.gov.scot
}
\details{
Sources list of available data sets from statistics.gov.uk and returns
  a data frame accounting for the available data sets, timespan and
  geographies.
}
\examples{
# Get list of themes with 'pupil' keyword
head(get_ToC(which = "pupil"), n = 5)
# Get full list
head(get_ToC(which = "pupil"), n = 5)
}
