% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_scotstat.R
\name{query_scotstat}
\alias{query_scotstat}
\title{Query Scotstat Repository}
\usage{
query_scotstat(query)
}
\arguments{
\item{query}{SPARQL query string}
}
\value{
A response object derived via \code{\link[httr]{POST}} function.
}
\description{
Send query to Scotat data.
}
\details{
Query Scotstat data using via \code{\link[httr]{POST}} function.
}
\examples{
\dontrun{query_scotstat(read_query_file(query_file("get_available_frameworks")))}


}
