% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmarterScotland.R
\docType{package}
\name{SmarterScotland-package}
\alias{SmarterScotland}
\alias{SmarterScotland-package}
\title{SmarterScotland: Sourcing data from opendata.gov.scot}
\description{
SmarterScotland provides a suite of convenience functions
  facilitating sourcing data from the
  \href{http://statistics.gov.scot}{statistics.gov.scot} portal.
}
\details{
For sample usage see \code{vignette("basic-usage", package = "SmarterScotland")}.
  For other usage, including charting and more advanced functionalities see
 \code{vignette("other-functions", package = "SmarterScotland")}.
}
\section{Acknowledgements}{

The following package is written in the open (both metaphorically and literally as I tend do undertake this development as a leisure reactivity while working on the Glasgow - Edinburgh commuter train), with big thanks to the whole Open Source community that contributed useful input through their questions, answers, discussions and comments via StackOverflow, GitHub and R-Help.
}

\section{Package options}{

\describe{
\item{\code{SmarterScotland.endpoint}}{Defaults to
  \code{http://statistics.gov.scot/sparql}. If need can be changed by
  modyfying \code{options(SmarterScotland.endpoint = "new_value")}. It's
  highly unlikely that you will ever need to change this option.}

\item{\code{SmarterScotland.response_check}}{Defaults to \code{warn}
  which results in warning if endpoint is not accessible. Other options are
  \code{stop} and \code{ignore} which respectively stop on errors or
  ignore checks.}
}
}

\author{
\itemize{
\item Maintainer/Author: Konrad Zdeb \email{konrad.zdeb@me.com}
}
}
