% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utility_functions}
\alias{utility_functions}
\alias{query_file}
\title{Handy utility functions}
\usage{
query_file(query_file_name)
}
\description{
Small and handy utility functions usually not useful outside
  the package enviornment.
}
\section{Functions}{
\itemize{
\item \code{query_file}: Wrapper for the \code{system.file} function
passing path to a given query file and taking care of potentially missing
extensions.
}}

\keyword{internal}
