% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/utils.R
\name{read_query_file}
\alias{read_query_file}
\alias{utility_functions}
\alias{query_file}
\title{Handy utility functions}
\usage{
read_query_file(file_path)

query_file(query_file_name)
}
\arguments{
\item{file_path}{File path string to source}

\item{query_file_name}{Name of query file}
}
\description{
Small and handy utility functions usually not useful outside
  the package enviornment.
}
\section{read_query_file}{
 C++ fast cleaner for query text file removing
  comments and line breaks.
}

\section{query_file}{
 Wrapper for the \code{system.file} function
  passing path to a given query file and taking care of potentially missing
  extensions.
}

